Add interviews admin interface for all states.
