+++ src/bootstrap.ts
@@ -63,6 +63,7 @@ export const bootstrapApp = (app: NestExpressApplication): void => {
     res.locals._ = _;
     res.locals.helpers = viewHelpers;
     res.locals.login = req.isAuthenticated();
+    res.locals.path = req.path;
     if (req.user) {
       res.locals.user = req.user;
       res.locals.userRole = req.user ? req.user.role : 'guest';
