--- views/interview/index.pug
@@ -1,26 +1,0 @@
-extends ../layout
-
-block content
-  .container.pt-5
-    h1.mb-4.text-center #{__('app.interview_requests')}
-    .table-responsive.table-sm.bg-white.shadow-sm
-      table.table
-        thead.thead-light
-          tr
-            th.p-3.border-0 #{__('interview.interviewee')}
-            th.p-3.border-0 #{__('interview.profession')}
-            th.p-3.border-0 #{__('interview.position')}
-            th.p-3.border-0 #{__('interview.description')}
-            if userRole === 'admin'
-              th.p-3.border-0= __('interview.actions')
-          tbody
-            if interviews.length > 0
-              each interview in interviews
-                tr
-                  td.p-3.align-middle.text-nowrap #{interview.interviewee.fullName}
-                  td.p-3.align-middle.text-nowrap #{interview.profession}
-                  td.p-3.align-middle.text-nowrap #{interview.position}
-                  td.p-3.align-middle #{interview.description}
-                  if userRole === 'admin'
-                    td.p-3.align-middle
-                      a.btn.btn-outline-primary.mt-3(href=`/interview/${interview.id}/assignment`)= __('interview.assignment.assign')
