+++ views/layout.pug
@@ -20,7 +20,7 @@ html(lang='ru')
         if userRole === 'admin'
           ul.navbar-nav.mr-auto
             li.nav-item
-              a.nav-link(href="/interview") #{__('app.interview_requests')}
+              a.nav-link(href="/interview/manage/application") #{__('app.manage_interviews')}
         ul.navbar-nav.ml-auto
           if login
             li.nav-item
